requirements.txt
anthropic>=0.92.0
pydantic>=2.0
python-dotenv>=1.0
pillow>=10.0
